requirements.txt
streamlit 
pandas 
matplotlib.pyplot 
seaborn 
plotly.express
time
numpy 


